server/requirements.txt
Django
django-filter
djangorestframework
elasticsearch-dsl
gunicorn
psycopg2-binary
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2